requirements.txt
opencv-python
piexif
datetime